backend/requirements-ml-stable.txt
# Stable ML stack for this project (Windows, Python 3.9/3.10)
# Use this file for model training to avoid TensorFlow/protobuf resolver breaks.

numpy==1.24.3
scipy==1.10.1
pandas==2.1.4
scikit-learn==1.3.2
opencv-python==4.8.1.78
Pillow==10.4.0

tensorflow==2.16.2
keras==3.3.3
tensorboard==2.16.2
protobuf==4.25.3
h5py==3.11.0

torch==2.3.1
torchvision==0.18.1
ultralytics==8.3.73
xgboost==2.1.3
joblib==1.3.2

# Useful for plots/evaluation scripts
matplotlib==3.8.4
seaborn==0.13.2


